Major changes to ProcessingHistory implementation and API. This implementation uses a multimap internally that will need to be rearranged for use with python, but provides a potentially compact and simple mechanism to store history data.   Cost is actually reconstructing the history chain will be more complicated, but idea is to keep ProcessingHistory, which is a component of working data object, as compact and simple as possible.   API changed for how a new processing stage was defined and the implementation of ProcessingHistory changed completely from previous - which I don't think would have ever worked.
Another major change was the elog was move out of Seismogram and
TimeSeries definitions to be part of ProcessingHistory.  That seemed
sensible because it allowed ProcessingHistory methods to post errors
to elog, which they would not have been able to do without some
evil constructs if elog was part of Seismogram and TimeSeries.
There are thus a series of changes related to Seismogram and TimeSeries
implementations.