Add snr functions (#275)
* Fix typo in percentage norm amplitude function

* Add binding code for EstimateBandwidth used for signal-to-noise
estimation functions.  Elected to move the EstimateBandwidth function
and related code to amplitudes section - was in deconvolution.

* Add BandwidthStatistics function.

* Add first draft of snr module

* Add pickle code binding code for MTPowerSpectrumEngine

* Add pickle binding code for several objects that need to have that capability

* Updates to documentation

* Add missing include guards

* Fix cmake file to compile snr.cc and then fix compilation errors in
snr.cc found when it actually was compiled

* Bug fix:  bundle_seed_data failed to set ensemble live - relic of conversion to LoggingEnsemble.

* Add binding for f_range method - missed in previous version

* Add upper frequency search start position parameter for EstimateBandwidth.  Whole string
of cascade changes where required for things using that function including the prototype
CNR3CDecon algorithm.  This version passes preliminary testing on real data, but
will need a little more polish before it can be pushed.

* change an int to size_t to silence compiler warning

* Add save_spectrum option to snr module and fix some obvious problems in PowerSpectrum pickle lambda

* Fix name change from PerfAmplitude (a typo) to PercAmplitude

* Fixed issues caused by missing Metadata inheritance

* Add a basic test for PowerSpectrum

* Fix mistake in not defining Metadata inheritance in the class template for PowerSpectrum

* Fixed typo in an error message.  Needs attention because there is
a design problem here.  The constructor prints a warning message
when the time bandwidth produce and number of tapers is inconsistent
and resets number of tapers to the maximum if int(2*tbp).
That is not the right behavior for mspass, but we don't want it to
throw an exception either.  May need a python wrapper to handle
this problem to avoid creating a ton of these messages if a user
makes an error.

* Add operator+= to ErrorLogger

* Fix mistake in previous commit - failed to remove scaffolding

* Lot of changes to make this algorithm better behaved

Co-authored-by: wangyinz <yinzhi.wang.cug@gmail.com>