Update snr.py
Fix a relatively harmless logic error in section of FD_snr_estimator that posts computed snr attributes to output python dictionary.   The error was harmless as attributes where posted but the logic created a pile of error log messages that were not correct - there was no such error.   Basically a set of if statements needed to be elif.