Update database.py
Simple fix to enable an archaic FIXME headed comment section handling of miniseed endtime.   A later revision on of the miniseed indexing C++ code now computes endtime correctly and properly handles data with data gaps.   There are other potential cascading errors in handling of gaps, but fixing this only changes what is contained in the wf_miniseed documents.  Without this the endtime and npts of a block of miniseed data is unknown.  Putting this back is essential for making the new time window reader work with continuous data. 

More work will be needed if this breaks any tests but if it does the tests are probably flawed.