Update reduce.py
Improve docstrings for dask and spark foldby operators.   The previous version had the dask and spark docstrings reversed.  I also added an example.  It took me a fair amount of trial and error to realize the usage in the example is correct.  Note I am only assuming the spark version is the same.