Update window.py
I ran across this code when I was trying to remember now to use the mspass_func_wrapper and noticed:
1. I does not have docstring
2. It is not correct as previous version of this had a bug in how it set the start of time window.   You can see the one line change I made.  It may break the tests.

I actually think this function is not needed.   To my mind it would be harder to remember a function with a similar but different name than to just remember what a TimeWindow object is or use a start and end time explicitly set as two parameters.