Update database.py
Fix typo.   BTW I found this because spyder flagged this as a syntax error.  It also flags lines 3083 and 3084 in the reader version of this pair.  I'm less sure how to fix that one.  Both were clearly not detected.   The one I fix here is in an error handler that is apparently not tested.   The other error is in the section for handling object level history which we've kind of pushed aside as secondary.