Fix ApplyCalib constructor bug
This fixes a rare bug for bad channel Metadata.  Hit an example where the other calibration data seeded valid but the sensitivity value was set to 0.0.   That caused a divide by 0 error exception when sensitivity is converted to calib (calib = 1/sensitivity).   This adds a small error handler that should take care of the problem.  Probably should have  pytest for this exception, but this is an rare error and indicates something got by IRIS’s QC procedures to have an invalid entry like that.