Fix ApplyCalib constructor bug (#633)
* Fix ApplyCalib constructor bug

This fixes a rare bug for bad channel Metadata.  Hit an example where the other calibration data seeded valid but the sensitivity value was set to 0.0.   That caused a divide by 0 error exception when sensitivity is converted to calib (calib = 1/sensitivity).   This adds a small error handler that should take care of the problem.  Probably should have  pytest for this exception, but this is an rare error and indicates something got by IRIS’s QC procedures to have an invalid entry like that.

* fix moto version to 5.1.1

* update cmake min version to 3.11

* force min version for yaml-cpp

* fall back to yaml-cpp 0.6.3

* let moto version to be less than 5.1.1

* downgrade pybind11 back to 2.10.1

* force pybind11 to newer min ver of cmake

---------

Co-authored-by: Ian Wang <22849821+wangyinz@users.noreply.github.com>
Co-authored-by: Ian Wang <yinzhi.wang.cug@gmail.com>